LPS-176632 Inline CompositeFacetProcessor into BaseSolrQueryAssemblerImpl, since it's only used there
